[DEV-4096] Refactoring
